+++ ts/features/walletV3/payment/saga/networking/handleWalletPaymentCalculateFees.ts
@@ -2,19 +2,14 @@ import * as E from "fp-ts/lib/Either";
 import { pipe } from "fp-ts/lib/function";
 import { call, put } from "typed-redux-saga/macro";
 import { ActionType } from "typesafe-actions";
-import { CalculateFeeRequest } from "../../../../../definitions/pagopa/ecommerce/CalculateFeeRequest";
-import { SagaCallReturnType } from "../../../../types/utils";
-import { getGenericError, getNetworkError } from "../../../../utils/errors";
-import { readablePrivacyReport } from "../../../../utils/reporters";
-import { withRefreshApiCall } from "../../../fastLogin/saga/utils";
-import { PaymentClient } from "../api/client";
-import { walletPaymentCalculateFees } from "../store/actions";
+import { CalculateFeeRequest } from "../../../../../../definitions/pagopa/ecommerce/CalculateFeeRequest";
+import { SagaCallReturnType } from "../../../../../types/utils";
+import { getGenericError, getNetworkError } from "../../../../../utils/errors";
+import { readablePrivacyReport } from "../../../../../utils/reporters";
+import { withRefreshApiCall } from "../../../../fastLogin/saga/utils";
+import { PaymentClient } from "../../api/client";
+import { walletPaymentCalculateFees } from "../../store/actions/networking";
 
-/**
- * Handle the remote call to get all available payment methods
- * @param getAllPaymentMethods
- * @param authToken
- */
 export function* handleWalletPaymentCalculateFees(
   calculateFees: PaymentClient["calculateFees"],
   action: ActionType<(typeof walletPaymentCalculateFees)["request"]>
